refactor(agent): split agent.rs into spawn/transcript/provider submodules
